+++ src/main/java/com/buuz135/industrial/proxy/CommonProxy.java
@@ -3,10 +3,12 @@ package com.buuz135.industrial.proxy;
 import com.buuz135.industrial.config.CustomConfiguration;
 import com.buuz135.industrial.utils.CraftingUtils;
 import com.buuz135.industrial.utils.RecipeUtils;
+import com.buuz135.industrial.utils.apihandlers.CraftTweakerHelper;
 import com.buuz135.industrial.utils.apihandlers.RecipeHandlers;
 import net.minecraft.util.DamageSource;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.config.Configuration;
+import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
 import java.util.Random;
@@ -23,6 +25,7 @@ public class CommonProxy {
         FluidsRegistry.registerFluids();
         RecipeHandlers.loadBioReactorEntries();
         RecipeHandlers.loadLaserLensEntries();
+        if (Loader.isModLoaded("crafttweaker")) CraftTweakerHelper.register();
 
         MinecraftForge.EVENT_BUS.register(new BlockRegistry());
         MinecraftForge.EVENT_BUS.register(new ItemRegistry());